Manage listings  added

+++ app/Http/Controllers/ListingController.php
@@ -5,28 +5,30 @@ namespace App\Http\Controllers;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Storage;
 
 class ListingController extends Controller
 {
-    //Show All Listing
+    // Show all listings
     public function index() {
         return view('listings.index', [
-            'listing' => Listing::latest()->filter(request(['tag', 'search']))->paginate(4)
+            'listings' => Listing::latest()->filter(request(['tag', 'search']))->paginate(6)
         ]);
     }
 
-    //Show Single Listing
+    //Show single listing
     public function show(Listing $listing) {
         return view('listings.show', [
             'listing' => $listing
         ]);
     }
-    //Show Create Form
+
+    // Show Create Form
     public function create() {
         return view('listings.create');
     }
 
-    //Store Listing Data
+    // Store Listing Data
     public function store(Request $request) {
         $formFields = $request->validate([
             'title' => 'required',
@@ -48,8 +50,19 @@ class ListingController extends Controller
 
         return redirect('/')->with('message', 'Listing created successfully!');
     }
-    //Update Listing Data
+
+    // Show Edit Form
+    public function edit(Listing $listing) {
+        return view('listings.edit', ['listing' => $listing]);
+    }
+
+    // Update Listing Data
     public function update(Request $request, Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -64,22 +77,27 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
-        $formFields['user_id'] = auth()->id();
-
         $listing->update($formFields);
 
         return back()->with('message', 'Listing updated successfully!');
     }
 
-    // Show Edit Form
-    public function edit (Listing $listing) {
-        return view('listings.edit', ['listing' => $listing]);
-    }
-
-    //Delete Listing
-    public function delete(Listing $listing) {
+    // Delete Listing
+    public function destroy(Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
+        if($listing->logo && Storage::disk('public')->exists($listing->logo)) {
+            Storage::disk('public')->delete($listing->logo);
+        }
         $listing->delete();
-    return redirect('/')->with('message', 'Listing deleted sucsessfully!');
+        return redirect('/')->with('message', 'Listing deleted successfully');
     }
 
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
+    }
 }
